Append video admin pages

+++ web/src/app/api/video/[id]/route.ts
@@ -1,6 +1,8 @@
 import { NextResponse } from "next/server";
 import dbConnect from '@/lib/mongoose';
 import Video from '@/lib/models/Video';
+import { getSession } from "@/lib/session";
+import { Role } from "@/lib/models/User";
 
 /**
  * @swagger
@@ -118,17 +120,21 @@ export async function PUT(request: Request, { params }: { params: { id: string }
  */
 export async function DELETE(request: Request, { params }: { params: { id: string } }) {
   try {
+    const session = await getSession();
+    if (!session || !session.isAuth) {
+      return NextResponse.json({ success: false, error: "Authentication required" }, { status: 401 });
+    }
+    if (session.role != Role.ADMIN) {
+      return NextResponse.json({ success: false, error: "Permission required" }, { status: 403 });
+    }
     await dbConnect();
     const videoId = params.id;
-
     const deletedVideo = await Video.findByIdAndDelete(videoId);
-
     if (!deletedVideo) {
-      return NextResponse.json({ success: false, message: "Video not found" }, { status: 404 });
+      return NextResponse.json({ success: false, error: "Video not found" }, { status: 404 });
     }
-
-    return NextResponse.json({ success: true, message: "Video deleted successfully" });
+    return NextResponse.json({ success: true });
   } catch (error: any) {
-    return NextResponse.json({ success: false, error: error.message }, { status: 500 });
+    return NextResponse.json({ success: false, error: "Internal server error" }, { status: 500 });
   }
 }
